move Landing component to generic layouts folder
